Implement startWith such that it can be called from any thread
The previous version called

    finalLiveData.value = startingValue

in the thread that calls startWith.  This is problematic, because setting the value of a LiveData
can only be done from the main thread.

This implementation avoids that and instead creates startWith as a merge of two LiveDatas: the
original source and a new live data from which only a single value is taken in case the source has
not emitted a value.  This is a bit more complicated than the original version, but allows startWith
to be called from any thread.

Unfortunately this is a breaking change in the sense that an attempt to get the value of the
returned LiveData will return null, because the initial value will not be set until the returned
LiveData is activated.