Allow null initial value in distinctUntilChanged
The previous implementation of distinctUntilChanged used a variable

    var latestValue : T? = null

and a check

    if(latestValue!=it)

to filter out duplicates.  This has the problem that an initial null value emitted by the source
will be implicitly dropped.

The new implementation uses a boolean variable to track whether the source has emitted anything.  In
case it hasn't the initial value will be emitted.

The new implementation also does not keep the previous value in a separate value.  Rather the
previous value is obtained from the returned live data.  This has the advantage that there is no
need to allocate space for a separate variable to track the previous value.